requirements.txt
Django==2.2.9
psycopg2-binary==2.8.4
djangorestframework==3.11.0
